Allow non-alphanumeric escapes
e.g.) :inf-Eg(A, [B\,C])
The premise would be parsed into `B\,C` although `B,C` was expected. This was
because the reader allowed a only `\]` escape when reading the premise. In
order to accept both `\]` and `\,` any `\X` escape (where X is not an
alphanumeric character) are allowed with this commit.